pull up CombineProjections
